+++ docs/CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ### Changed
 - Show "Pressure (Pa)" or "Height (m)" in plot labels instead of just "Level"
-- Vertical profile plots now show zero-pressure level at the top
+- "Vertical Profile" plots have been renamed to "Station Profile"
+    - Also, they now show zero-pressure level at the top
 
 ### Fixed
 - Bug in query stub which would cause "Average Analysis Departure Map" plots